Refactored serializer and made it generic
